Remove component CatalogMain and relocate all code to page Catalog
